Fix avg_profit_pct not being saved - cast AVG() to float8
The get_raw_profit_pct() query was returning numeric type from PostgreSQL AVG(),
but sqlx was trying to read it as f64. This caused .try_get() to fail silently,
returning None which unwrapped to 0.0.

Added ::float8 cast to match the pattern used in get_raw_win_rate().

Result: Wallet profiles will now correctly show avg_profit_per_trade_pct values
like 5.5746 (557% profit) instead of 0.0000.

+++ orderflow-reputation/src/calculator.rs
@@ -332,7 +332,7 @@ impl ReputationCalculator {
                     AND t2.side = 'SELL'
                     AND t2.timestamp > t1.timestamp
             )
-            SELECT AVG(profit_pct) as avg_profit_pct
+            SELECT AVG(profit_pct)::float8 as avg_profit_pct
             FROM position_pairs
             "#
         )
